refactor(cli): split serve server routes
